Single player gameplay
This commit adds the turn mechanic so the user can take a turn at
selecting a enemy tile and the enemy then gets a go at selecting a
player tile.

This does not implement the gameover system or the ability to sink enemy
ships as the enemy board is not yet populated with ships to hit.

- Methods added to the GameManager script to enable to turn based
mechanic
- Refactored the code for player and enemy logic
- Implemented a delay for the player and enemy moves so the user can
have a chance to see what is happening
- Changed the music track to the dark synthwave which plays when the
setup scene loads and continues during gameplay
Made tweaks to the EnemyTile so it renders the same size as the neon
tiles