Single player game improvements following initial implementation
- Prevented the user from having more than one go on the enemy board by
adding a check to only execute the turn logic if it is the player's
turn.  Also amended the code so the turn is changed as soon as the
player turn is displayed.
- Fixed edge case where user could click on their own board and place
another ship if the enemy sunk a ship.  This was because when the player
ship count went below the max limit, the code allowed ship placemet
again.  I added a check so that if the game has started, ship placement
is disabled.
- Fixed another edge case in the enemy board, where users could click on squares they had alread clicked on their previous go.  I've added an additional check to see if there is a child and whether the child has the 'miss' tag.  If so, the turn logic is not executed, preventing the miss being replaced with a hit and the player messing up their turn.
- Also addressed an error that had been appearing regarding the use of
serialized fields with the same name in either parent or child.  I've
made the Enemy Tile inherit the Tile's gameManager so it doesn't have to
create it's own reference to the gameManager.  This seems to have
resolved the issue but not sure if this is right.  Added a comment so
this can be looked at again.