Single player gameplay elements completed
- Added logic to instantiate 5 enemy ships and add individual ships as
children of 5 random EnemyTile gameObjects
- Improved the GetRandomNumber method to make it reusable
- Tweaked the player logic so that when a player sinks a ship, the ship
is set to active so the user can see the ship as well as the red hit
marker
-- Uncommented the decrease enemy ship count line in the
DecrementShipList method
Created a CheckGameOver method to check if the player had one, lost or
if the game should continue.  If the game continued, ToggleTurn is
called.  If the player wins, PlayerWin is called to stop the game and
displaye the appropriate message.  If enemy wins, PlayerLose is called,
the game is stopped and the appropriate message is displayed
- Added logic to prevent the player clicking on a tile where they have
already sunk a ship.  This prevents the player being able to lower the
enemy ship count artifically.
- Fixed the transitions between turns
- Added titles for the boards that change between turns
- Added a title for the setup section of the game
- Updated the Board script so the boards are centered
- Now includes new and improved audio track exclusive to the super delux
ultra elite premium edition