This commit implements functionality to play again, return to the main menu and also improves the presentation.
- Updated GameManager and SinglePlayerSetupCanvas so they are no longer
using the singleton pattern.  I found that persisting them with our
current game setup caused them to be duplicated when reloading a scene
to start a new game.  The code has just been commented out for now in
case we decide to go with separate scenes after all.
- Created a panel for the Game over messages and added the buttons for returning
to the main menu and starting a new game.  Also added code to set the
relvant game over message to active depending on whether they player won
or not.
- Tweaked the game transition code so there isn't a delay in the
transition when the user clicks play.
Added new assets for the panel background and images to use for the
titles and game over messages.
- Tweaked the settings for the canvas scaler on the
SinglePlayerSetupCanvas so the titles scale.
- Updated the anchors for the titles, panel and play button, to keep
them in the right areas when the screen size changes.